+++ client/src/components/Chat/ConversationMessage/index.tsx
@@ -95,6 +95,7 @@ const ConversationMessage = ({
                 <FileChip
                   onClick={() => openFileModal(s.path)}
                   fileName={s.path.split('/').pop() || ''}
+                  filePath={s.path || ''}
                 />
               ) : null}
             </div>
