+++ client/src/pages/ArticleResponse/index.tsx
@@ -1,20 +1,11 @@
-import { useContext, useEffect, useMemo, useState } from 'react';
-import { useTranslation } from 'react-i18next';
-import { conversationsCache } from '../../services/cache';
-import {
-  ChatMessage,
-  ChatMessageAuthor,
-  ChatMessageServer,
-} from '../../types/general';
-import { ChatContext } from '../../context/chatContext';
+import { useContext, useState } from 'react';
 import { FileModalContext } from '../../context/fileModalContext';
 import { UIContext } from '../../context/uiContext';
-import { getConversation } from '../../services/api';
-import { mapLoadingSteps } from '../../mappers/conversation';
 import MarkdownWithCode from '../../components/MarkdownWithCode';
 import Button from '../../components/Button';
 import { CopyMD } from '../../icons';
 import { copyToClipboard } from '../../utils';
+import useConversation from '../../hooks/useConversation';
 
 type Props = {
   recordId: number;
@@ -22,9 +13,6 @@ type Props = {
 };
 
 const ArticleResponse = ({ recordId, threadId }: Props) => {
-  const { t } = useTranslation();
-  const { conversation } = useContext(ChatContext.Values);
-  const { setThreadId, setConversation } = useContext(ChatContext.Setters);
   const { openFileModal } = useContext(FileModalContext);
   const { tab } = useContext(UIContext.Tab);
   const [data, setData] = useState(
@@ -44,35 +32,7 @@ const ArticleResponse = ({ recordId, threadId }: Props) => {
     threadId,
   ]);
 
-  useEffect(() => {
-    if (!conversationsCache[threadId]?.[recordId] && !conversation[recordId]) {
-      getConversation(threadId).then((resp) => {
-        const conv: ChatMessage[] = [];
-        resp.forEach((m) => {
-          // @ts-ignore
-          const userQuery = m.search_steps.find((s) => s.type === 'QUERY');
-          conv.push({
-            author: ChatMessageAuthor.User,
-            text: m.query?.target?.Plain || userQuery?.content?.query || '',
-            isFromHistory: true,
-          });
-          conv.push({
-            author: ChatMessageAuthor.Server,
-            isLoading: false,
-            loadingSteps: mapLoadingSteps(m.search_steps, t),
-            text: m.conclusion,
-            results: m.answer,
-            isFromHistory: true,
-            queryId: m.id,
-            responseTimestamp: m.response_timestamp,
-          });
-        });
-        conversationsCache[threadId] = conv;
-        setThreadId(threadId);
-        setConversation(conv);
-      });
-    }
-  }, []);
+  const { data } = useConversation(threadId, recordId);
 
   return (
     <div className="overflow-auto p-8 w-screen">
